feat: 사용자 정보 관리 및 로그인 상태 처리 개선
