+++ lib/presentation/splash/view_models/splash_view_model.dart
@@ -1,19 +1,18 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/remote_user_info_data_source.dart';
-import 'package:mongbi_app/data/dtos/user_dto.dart';
+import 'package:mongbi_app/domain/entities/user.dart';
+import 'package:mongbi_app/domain/use_cases/get_user_info_use_case.dart';
+import 'package:mongbi_app/presentation/remind/view_model/remind_time_setting_view_model.dart';
 import 'package:mongbi_app/presentation/splash/view_models/splash_state.dart';
 import 'package:mongbi_app/presentation/splash/view_models/splash_status.dart';
+import 'package:mongbi_app/presentation/auth/viewmodels/auth_view_model.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-
-import 'package:mongbi_app/providers/user_info_provider.dart';
-import 'package:mongbi_app/presentation/remind/view_model/remind_time_setting_view_model.dart';
 
 
 class SplashViewModel extends StateNotifier<SplashState> {
-  final RemoteUserInfoGetDataSource _dataSource;
-  final Ref ref;
 
-  SplashViewModel(this._dataSource, this.ref) : super(const SplashState());
+  SplashViewModel(this._getUserInfoUseCase, this.ref) : super(const SplashState());
+  final GetUserInfoUseCase _getUserInfoUseCase;
+  final Ref ref;
 
   Future<void> checkLoginAndFetchUserInfo() async {
     state = state.copyWith(status: SplashStatus.loading);
@@ -27,10 +26,8 @@ class SplashViewModel extends StateNotifier<SplashState> {
         return;
       }
 
-      final userList = await _dataSource.fetchGetUserInfo();
-      ref.read(currentUserProvider.notifier).setUser(userList[0].toEntity());
-
-      // Check for reminder time
+      final userList = await _getUserInfoUseCase.execute();
+      ref.read(currentUserProvider.notifier).setUser(userList[0]);
       final reminderTime = await NotificationService().loadReminderTime();
       if (reminderTime != null) {
         state = state.copyWith(status: SplashStatus.successWithReminder, userList: userList);
@@ -42,7 +39,7 @@ class SplashViewModel extends StateNotifier<SplashState> {
     }
   }
 
-  void setUser(UserDto user) {
+  void setUser(User user) {
     state = state.copyWith(status: SplashStatus.success, userList: [user]);
   }
