+++ lib/providers/user_info_provider.dart
@@ -31,21 +31,8 @@ final userInfoDataSourceProvider = Provider<RemoteUserInfoGetDataSource>(
 
 final splashViewModelProvider =
     StateNotifierProvider<SplashViewModel, SplashState>((ref) {
-      final dataSource = ref.read(userInfoDataSourceProvider);
-      return SplashViewModel(dataSource, ref);
+      ref.read(userInfoDataSourceProvider);
+      return SplashViewModel(ref.read(getUserInfoUseCaseProvider), ref);
     });
 
-final currentUserProvider = NotifierProvider<CurrentUserNotifier, User?>(
-  () => CurrentUserNotifier(),
-);
 
-class CurrentUserNotifier extends Notifier<User?> {
-  @override
-  User? build() {
-    return null; // Initial state is null (no user logged in)
-  }
-
-  void setUser(User? user) {
-    state = user;
-  }
-}
